Rename src/binding.* to src/tf_binding.x
