+++ tools/build.js
@@ -6,6 +6,6 @@ if (process.argv.indexOf("clean") >= 0) {
 }
 
 (async() => {
-  run.sh("node ./tools/build_binding.js");
+  run.sh("node ./tools/build_tf_binding.js");
   run.sh("node ./tools/website_render.js");
 })();
