+++ tools/cpplint.js
@@ -1,3 +1,3 @@
 #!/usr/bin/env node
 const run = require("./run");
-run.sh("python ./deps/cpplint/cpplint.py src/binding.cc src/check.h");
+run.sh("python ./deps/cpplint/cpplint.py src/tf_binding.cc src/check.h");
